\name{biomassVector}
\alias{biomassVector}
\title{
The creation of the biomass vector required for packing into the enaR network object
}
\description{
  The generation of the biomass vector required for use in packing a lim file into an enaR network object.
}
\usage{
biomassVector(limfile)
}
\arguments{
  \item{limfile}{
This is the user generated LIM file created for analysis purposes.

}
}
\author{
Ruchit Mahabir

Christopher Waspe

Ursula Scharler}

\examples{

biomassVecList <- biomassVector(limfile = N4)

}

