\name{FlowCAr}
\alias{FlowCAr}
\title{
Flow Network Construction and Analysis
}
\description{
This package allows for the creation of multiple possible network models based on a single FlowCAr input file.
}
\details{
The following table indicates the functions used in the FlowCAr package.
\tabular{ll}{
Procedure \tab FlowCAr Functions\cr
     -\tab - \cr
Network Construction \tab \code{\link{LIMbuild}}\cr
      \tab \code{\link{limListGen}}\cr
      \tab \code{\link{flowCheck}}\cr
      \tab \code{\link{flowLimit}}\cr
Restructuring and Packing \tab \code{\link{inputVector}}\cr
      \tab \code{\link{exportVector}}\cr
      \tab \code{\link{respVector}}\cr
      \tab \code{\link{outputVector}}\cr
      \tab \code{\link{biomassVector}}\cr
      \tab \code{\link{livingVector}} \cr
      \tab \code{\link{PackNet}}\cr
Main Function \tab \strong{\code{\link{enaListGen}}}\cr
Network Validation and Visualisation  \tab \code{\link{plotNodeFlows}}\cr
      \tab \code{\link{plotRange}}\cr
      \tab \code{\link{plotFlowRange}}\cr
}
}
\references{
     
     #TBC
     
     }
\author{Ruchit Mahabir

Christopher Waspe

Ursula Scharler
}
