\name{N4LIM}
\alias{N4LIM}
\docType{data}
\title{
Solved Four Node Network
}
\description{
Output from limListGen run on N4. 1000 iterations. NULL jumps.
}

\format{
 A FlowCAr output containing a list of 2 variables.
  
}

\source{
FlowCAr package
}
\references{
FlowCAr ref
}

\keyword{datasets}
