\name{N4LIMv}
\alias{N4LIMv}
\docType{data}
\title{
Restricted Solved Four Node Network
}
\description{
Output from flowLimit run on N4LIM, restricted Invertebrate flow.
}

\format{
 A FlowCAr output containing a list of 2 variables.
  
}

\source{
FlowCAr package
}
\references{
FlowCAr ref
}

\keyword{datasets}
