\name{N4list}
\alias{N4list}
\docType{data}
\title{
Solved and packed four node network
}
\description{
Output from FlowCAr main function, enaListGen run on "4node.R". 1000 iterations. NULL jumps.
}

\format{
 A FlowCAr output containing a list of 3 variables.
  
}

\source{
FlowCAr package
}
\references{
FlowCAr ref
}

\keyword{datasets}
