\name{biomassVector}
\alias{biomassVector}
\title{
The creation of the biomass vector required for packing into the enaR network object
}
\description{
}
\usage{
biomassVector(limfile)
}
\arguments{
  \item{limfile}{
This is the user generated LIM file created for analysis purposes.

}
}
\details{
}
\value{

}
\references{
}
\author{
Ruchit Mahabir

Christopher Waspe

Ursula Scharler}
\note{
}

\seealso{
}
\examples{

biomassVecList <- biomassVector(limfile = N4)

}

\keyword{ ~kwd1 }
\keyword{ ~kwd2 }
